Machamp Vital Throw — depend on real IKA source, kill the shim
Switch silvally/Move.toml from the local ika_shim to the real IKA Move
source at dwallet-labs/ika@mainnet-v1.1.8 (deployed_contracts/mainnet
subdir — the one with [env.mainnet] published-ids baked in).

The shim broke Sui 1.69.2's upgrade-graph resolution with
PublishUpgradeMissingDependency. The real source fixes that AND
surfaced two real signature mismatches the shim had masked:
 - approve_message takes signature_algorithm: u32, hash_scheme: u32
   (shim had u8/u8)
 - MessageApproval lives in coordinator_inner, not coordinator

Delete ika_shim/ entirely. No more manual bytecode disassembly
required — IKA publishes their Move source.

Follow-up: Machamp Seismic Toss (mainnet publish) lands in the next
commit with the Published.toml pinning the package id.

+++ contracts/silvally/sources/dwallet_subname_policy.move
@@ -20,8 +20,8 @@
 
 module ski::dwallet_subname_policy;
 
-use ika_dwallet_2pc_mpc::coordinator::{Self, DWalletCoordinator, MessageApproval};
-use ika_dwallet_2pc_mpc::coordinator_inner::DWalletCap;
+use ika_dwallet_2pc_mpc::coordinator::{Self, DWalletCoordinator};
+use ika_dwallet_2pc_mpc::coordinator_inner::{DWalletCap, MessageApproval};
 use sui::clock::Clock;
 use sui::event;
 
@@ -43,8 +43,8 @@ const E_NOT_OWNER: u64 = 3;
 // Curve is implied by the DWalletCap (set at DKG time).
 // ------------------------------------------------------------------
 
-const SIG_ALG_ECDSA_SECP256K1: u8 = 0;
-const HASH_KECCAK256: u8 = 0;
+const SIG_ALG_ECDSA_SECP256K1: u32 = 0;
+const HASH_KECCAK256: u32 = 0;
 
 // ------------------------------------------------------------------
 // State
